Missed implementations

+++ Usings/GlobalUsings.cs
@@ -27,5 +27,5 @@ global using ShadowDecalHandle_t = ushort;
 global using OverlayFragmentHandle_t = ushort;
 global using Trace = Source.Common.GameTrace;
 global using ViewRects = Source.Common.Utilities.RefStack<Source.Common.Engine.ViewRect>;
-global using ClientRenderHandle_t = nint;
+global using ClientRenderHandle_t = long;
 global using static Source.Common.ClientRenderHandleGlobals;
